+++ hw/usb/desc.h
@@ -239,6 +239,6 @@ int usb_desc_string(USBDevice *dev, int index, uint8_t *dest, size_t len);
 int usb_desc_get_descriptor(USBDevice *dev, USBPacket *p,
         int value, uint8_t *dest, size_t len);
 int usb_desc_handle_control(USBDevice *dev, USBPacket *p,
-        int request, int value, int index, int length, uint8_t *data);
+        int request, int value, int index, size_t length, uint8_t *data);
 
 #endif /* QEMU_HW_USB_DESC_H */
